# api.yaml
openapi: 3.0.3
info:
  title: Room Booking Service (Test Task)
  version: "1.0.0"
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
    Token:
      type: object
      properties:
        token:
          type: string
    Room:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string, nullable: true }
        capacity: { type: integer, nullable: true }
        createdAt: { type: string, format: date-time, nullable: true }
    Schedule:
      type: object
      properties:
        id: { type: string, format: uuid }
        roomId: { type: string, format: uuid }
        daysOfWeek:
          type: array
          items: { type: integer }
        startTime: { type: string }
        endTime: { type: string }
    Slot:
      type: object
      properties:
        id: { type: string, format: uuid }
        roomId: { type: string, format: uuid }
        start: { type: string, format: date-time }
        end: { type: string, format: date-time }
    Booking:
      type: object
      properties:
        id: { type: string, format: uuid }
        slotId: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        status: { type: string, enum: [active, cancelled] }
        conferenceLink: { type: string, nullable: true }
        createdAt: { type: string, format: date-time, nullable: true }
paths:
  /_info:
    get:
      summary: health endpoint
      responses:
        '200':
          description: ok
  /dummyLogin:
    post:
      summary: get dummy token by role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  type: string
                  enum: [admin, user]
      responses:
        '200':
          description: token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
  /register:
    post:
      summary: register user (optional)
      responses:
        '201': { description: created }
  /login:
    post:
      summary: login (optional)
      responses:
        '200':
          description: token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
  /rooms/list:
    get:
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: rooms }
  /rooms/create:
    post:
      security: [{ BearerAuth: [] }]
      responses:
        '201': { description: room created }
  /rooms/{roomId}/schedule/create:
    post:
      security: [{ BearerAuth: [] }]
      parameters:
        - name: roomId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '201': { description: schedule created }
  /rooms/{roomId}/slots/list:
    get:
      security: [{ BearerAuth: [] }]
      parameters:
        - name: roomId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: date
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        '200': { description: available slots }
  /bookings/create:
    post:
      security: [{ BearerAuth: [] }]
      responses:
        '201': { description: booking created }
  /bookings/list:
    get:
      security: [{ BearerAuth: [] }]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        '200': { description: bookings list }
  /bookings/my:
    get:
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: user bookings }
  /bookings/{bookingId}/cancel:
    post:
      security: [{ BearerAuth: [] }]
      parameters:
        - name: bookingId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: cancelled or already cancelled }